v0.2.5: fixed install and typing problems, benchmark improvements

+++ benchmark/runner.ts
@@ -1,10 +1,11 @@
 import { fork } from 'child_process';
-import { fileURLToPath } from 'url';
-import * as path from 'path';
 import * as fs from 'fs';
 import * as olmdb from 'olmdb';
+import * as path from 'path';
+import { fileURLToPath } from 'url';
 
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
+// @ts-ignore
+if (typeof __dirname === 'undefined') global.__dirname = (typeof import.meta === 'undefined') ? process.cwd() : path.dirname(fileURLToPath(import.meta.url));
 
 function sleep(ms: number) {
     return new Promise(resolve => setTimeout(resolve, ms));
@@ -22,22 +23,28 @@ function parseArgs() {
         db_dir: "./.olmdb_benchmark",
         duration: 10, // seconds
         delay: 0,
+        keep: false,
     };
 
     for (let i = 0; i < args.length; i++) {
         const arg = args[i];
         if (arg.startsWith('--')) {
-            let [key, value]: (string|number)[] = arg.substring(2).split('=');
+            let [key, value]: (string|number|boolean)[] = arg.substring(2).split('=');
             key = key.replace(/-/g, '_'); // Convert dashes to underscores
             if (key in options) {
-                value ||= args[++i];
-                if (typeof options[key] === 'number') value = parseInt(value, 10);
+                if (typeof options[key] === 'boolean') {
+                    value = (value !== 'false' && value !== '0' && value !== 'no');
+                } else {
+                    value ||= args[++i];
+                    if (typeof options[key] === 'number') value = parseInt(value, 10);
+                }
                 options[key] = value;
-            } else {
-                console.error(`Unknown option: ${key}`);
-                process.exit(1);
+                continue;
             }
         }
+        console.error(`Unknown option: ${arg}`);
+        process.exit(1);
+
     }
     return options;
 }
@@ -46,23 +53,27 @@ async function main() {
     const options = parseArgs();
 
     // Delete benchmark database directory
-    if (fs.existsSync(options.db_dir)) {
-        fs.rmSync(options.db_dir, { recursive: true, force: true });
+    if (options.key_count >= 0) {
+        if (fs.existsSync(options.db_dir)) {
+            fs.rmSync(options.db_dir, { recursive: true, force: true });
+        }
     }
-
     // Open database and create initial data
     olmdb.init(options.db_dir);
-    console.warn('Populating database...');
-    const value = new Uint8Array(options.value_size).fill('i'.charCodeAt(0));
-    let cnt = 0;
-    while(cnt < options.key_count) {
-        // Max 10000 keys per transaction
-        await olmdb.transact(() => {
-            console.warn(`Adding keys ${cnt} to ${cnt + 10000}...`);
-            for (let i = 0; i < 10000 && cnt < options.key_count; i++, cnt++) {
-                olmdb.put(cnt.toString(), value);
-            }
-        });
+
+    if (options.key_count >= 0) {
+        console.warn('Populating database...');
+        const value = new Uint8Array(options.value_size).fill('i'.charCodeAt(0));
+        let cnt = 0;
+        while(cnt < options.key_count) {
+            // Max 10000 keys per transaction
+            await olmdb.transact(() => {
+                console.warn(`Adding keys ${cnt} to ${cnt + 10000}...`);
+                for (let i = 0; i < 10000 && cnt < options.key_count; i++, cnt++) {
+                    olmdb.put(cnt.toString(), value);
+                }
+            });
+        }
     }
 
     if (options.delay > 0) {
@@ -77,6 +88,7 @@ async function main() {
     const workerPath = path.resolve(__dirname, 'worker.ts');
     if (options.threads > 1) {
         for (let i = 0; i < options.threads; i++) {
+            const seed = (i * 0xdeadbeef) ^ i;
             const worker = fork(workerPath);
             resultPromises.push(new Promise((resolve, reject) => {
                 worker.on('message', (result: any) => {
@@ -90,13 +102,13 @@ async function main() {
                         reject(new Error(`Worker stopped with exit code ${code}`));
                     }
                 });
-                worker.send(options);
+                worker.send({...options, seed});
             }));
         }
         await Promise.all(resultPromises);
     } else {
         const {run} = await import(workerPath);
-        totals = await run(options);
+        totals = await run({...options, seed: 0});
     }
 
     console.log(JSON.stringify(totals));
